refactor: Move official weights test to integration tests
- Create new tests/integration/models/ directory
- Move test_kimi_vl_official_weights to tests/integration/models/test_kimi_vl.py
- Remove integration test from main unit test suite
- Clean up unused imports from main test file

This separates slow integration tests that require external resources
from fast unit tests, keeping the main test suite focused and quick.
